+++ NEWS.md
@@ -1,6 +1,7 @@
 # TreeTools 0.1.1.9000 (development)
 
-- `RenumberTips` can extract tip order from `phylo` and `Splits` objects. 
+- `RenumberTips` can extract tip order from `phylo` and `Splits` objects.
+- Convert numeric tip labels to class `character` in `GetTipNames`.
 
 # TreeTools 0.1.1